Optimum conditions for pesticide application

Determining the ideal time for pesticide application using ∆T

deltat(
  LON,
  LAT,
  type = 2,
  days = 7,
  control = NULL,
  details = FALSE,
  dates = NULL,
  plot = FALSE
)

Arguments

LON

Longitude (in decimal)

LAT

Latitude (in decimal)

type

Type of analysis. Use 1 for forecast and 2 for temporal data.

days

Number of days (only use this argument if type=1).

control

Type of product to be applied. Use “fung” for fungicide, “herb” for herbicide, “ins” for insecticides, “bio” for biological products.

details

Returns the result in detail if TRUE.

dates

Only use this argument if type=2. Start and end date for obtaining weather data for a crop cycle.

Value

Returns the ideal application times, considering each scenario. Taking as a parameter a DELTA_T between 2 and 8, wind speed between 3 and 8, and no precipitation.

Author

Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva

Examples

# \donttest{
library(Breeding)
#> 
#> Attaching package: 'Breeding'
#> The following objects are masked from 'package:EstimateBreed':
#> 
#>     ALELICA, Coeficiente_endogamia, GS, GS2, GS3, GS4, Jinks_Pooni,
#>     SEGREGAÇÃO_PADRÃO, TEMP_BASE, blupis, cvar, deltat, desv_clim,
#>     estresse, fototermal, genpar, gga, gvri, heterose, iaf,
#>     index_vigor, indger, indviab, is_ptnerg, is_qindustrial, isgr, itu,
#>     ivig_simp, linearest, mut_index, ph, plastocrono, reg_GP, rend_ind,
#>     restr, risco, somatermica, speedb, tamef, tindex, transgressivos
#> The following object is masked from 'package:stats':
#> 
#>     deltat

# Forecasting application conditions
deltat(-53.696944444444,-28.063888888889,type=1,days=10)
#> Loading required package: httr
#> Loading required package: jsonlite
#> Loading required package: dplyr
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union
#> Loading required package: lubridate
#> 
#> Attaching package: 'lubridate'
#> The following objects are masked from 'package:base':
#> 
#>     date, intersect, setdiff, union
#> Momentos com condição ideal de aplicação
#> --------------------------------------------------------
#>           Dia  Hora Temp UR WindS Prec   DELTAT
#> 1  2025-02-16 23:00 23.6 88   9.4  0.0 2.109026
#> 2  2025-02-17 04:00 23.1 88   7.3  0.1 2.100995
#> 3  2025-02-17 05:00 24.5 84   9.4  0.3 2.887783
#> 4  2025-02-17 06:00 26.3 80   8.2  0.2 3.734931
#> 5  2025-02-17 13:00 27.2 80   9.8  1.4 3.760259
#> 6  2025-02-18 06:00 20.6 87   8.9  0.0 2.243733
#> 7  2025-02-18 17:00 23.3 68   6.2  0.0 6.246637
#> 8  2025-02-18 18:00 22.5 71   8.8  0.0 5.528588
#> 9  2025-02-18 23:00 19.9 82   8.1  0.0 3.168462
#> 10 2025-02-19 00:00 19.3 85   6.6  0.0 2.588524
#> 11 2025-02-19 01:00 18.7 88   4.5  0.0 2.030977
#> 12 2025-02-19 05:00 19.5 87   3.3  0.0 2.224745
#> 13 2025-02-19 06:00 20.8 84   1.1  0.0 2.807293
#> 14 2025-02-19 07:00 24.0 74   2.3  0.0 4.928860
#> 15 2025-02-19 08:00 25.9 65   3.4  0.0 7.095410
#> 16 2025-02-19 16:00 25.9 63   3.7  0.0 7.595337
#> 17 2025-02-19 17:00 23.8 74   5.1  0.0 4.921400
#> 18 2025-02-19 18:00 22.8 81   5.2  0.0 3.437423
#> 19 2025-02-19 19:00 22.2 85   4.7  0.0 2.646971
#> 20 2025-02-20 06:00 23.8 80   7.1  0.0 3.665020
#> 21 2025-02-20 07:00 25.6 71   5.2  0.0 5.659675
#> 22 2025-02-20 18:00 25.1 63   3.7  0.0 7.549965
#> 23 2025-02-20 19:00 23.9 68   3.4  0.0 6.275045
#> 24 2025-02-20 20:00 23.0 72   3.6  0.0 5.327631
#> 25 2025-02-20 21:00 22.4 76   4.2  0.0 4.445440
#> 26 2025-02-20 22:00 22.0 80   4.1  0.0 3.615087
#> 27 2025-02-20 23:00 21.9 84   3.5  0.0 2.831105
#> 28 2025-02-21 00:00 21.7 87   2.9  0.0 2.262800
#> 29 2025-02-21 05:00 24.3 81   3.1  0.0 3.476859
#> 30 2025-02-21 06:00 26.5 73   2.7  0.0 5.245020
#> 31 2025-02-21 07:00 27.9 67   1.6  0.0 6.708155
#> 32 2025-02-21 17:00 27.2 69   3.3  0.0 6.194616
#> 33 2025-02-21 18:00 25.7 75   4.0  0.0 4.773966
#> 34 2025-02-21 19:00 24.7 79   3.7  0.0 3.895040
#> 35 2025-02-21 20:00 23.9 82   3.2  0.0 3.267001
#> 36 2025-02-21 21:00 23.3 84   2.8  0.0 2.861555
#> 37 2025-02-21 22:00 22.8 86   2.6  0.0 2.469587
#> 38 2025-02-21 23:00 22.4 87   2.5  0.0 2.274976
#> 39 2025-02-22 00:00 22.1 87   2.9  0.0 2.269754
#> 40 2025-02-22 01:00 21.7 88   2.9  0.0 2.078589
#> 41 2025-02-22 02:00 21.4 88   2.7  0.0 2.073803
#> 42 2025-02-22 03:00 21.8 86   3.1  0.0 2.450752
#> 43 2025-02-22 04:00 23.4 80   3.9  0.0 3.653895
#> 44 2025-02-22 05:00 25.7 72   5.2  0.0 5.437406
#> 45 2025-02-22 06:00 27.7 65   5.4  0.0 7.191322
#> 46 2025-02-22 17:00 27.2 71   5.1  0.3 5.727920
#> 47 2025-02-22 18:00 25.4 80   5.2  0.3 3.709688
#> 48 2025-02-22 19:00 24.4 84   4.8  0.0 2.885593
#> 49 2025-02-22 20:00 23.9 86   4.3  0.0 2.490388
#> 50 2025-02-22 21:00 23.4 88   3.9  0.0 2.105812
#> 51 2025-02-23 04:00 22.6 88   6.1  0.0 2.092979
#> 52 2025-02-23 05:00 24.6 81   7.6  0.0 3.484773
#> 53 2025-02-23 06:00 26.6 74   8.4  0.0 5.026343
#> 54 2025-02-23 07:00 28.6 66   7.9  0.0 6.989669
#> 55 2025-02-23 18:00 27.1 62   4.5  0.0 7.921097
#> 56 2025-02-23 19:00 25.9 68   4.6  0.0 6.370193
#> 57 2025-02-23 20:00 24.9 72   3.9  0.0 5.404766
#> 58 2025-02-23 21:00 24.2 76   3.7  0.0 4.506649
#> 59 2025-02-23 22:00 23.6 78   3.7  0.0 4.068192
#> 60 2025-02-23 23:00 23.2 80   3.4  0.0 3.648338
#> 61 2025-02-24 00:00 22.8 81   3.6  0.0 3.437423
#> 62 2025-02-24 01:00 22.2 83   3.8  0.0 3.030257
#> 63 2025-02-24 02:00 21.7 86   4.1  0.0 2.448873
#> 64 2025-02-24 03:00 21.9 85   4.8  0.0 2.640895
#> 65 2025-02-24 04:00 23.4 80   6.6  0.0 3.653895
#> 66 2025-02-24 05:00 25.8 72   8.9  0.0 5.441492
#> 67 2025-02-26 18:00 20.4 88   9.4  0.1 2.057890
View(forecast)

# Retrospective analysis of application conditions
deltat(-53.696944444444,-28.063888888889,type=2,days=10,dates=c("2023-01-01","2023-05-01"))
#> Loading required package: nasapower
View(retrospective)
# }