Stress indices for genotype selection

Selection indices for genotypes conducted under stress conditions cited by Ghazvini et al. (2024).

estresse(
  GEN,
  YS,
  YC,
  index = "ALL",
  bygen = T,
  plot = F,
  xlab = "Genotype",
  ylab = "Values",
  ...
)

Arguments

GEN

The column with the genotypes to be selected.

YS

Productivity of the genotype without stress conditions.

YC

Genotype productivity under stressful conditions.

index

Index to be calculated (Standard “ALL”). The indices to be used are: 'STI' - Stress Tolerance Index, 'YI' - Yield Index, 'GMP' - Geometric Mean Productivity, 'MP' - Mean Productivity, 'MH' - Harmonic Mean, 'SSI' - Stress Stability Index, 'YSI' - Yield Stability Index, 'RSI' - Relative Stress Index.

bygen

Returns the average of each genotype if “TRUE”. Only in this way it will be possible to plot graphs.

plot

Plot graph if equal to “TRUE” (Standard “F”).

xlab

Adjust the title of the x-axis in the graph.

ylab

Adjust the title of the y-axis in the graph.

Value

Returns a table with the genotypes and the selected indices. The higher the index value, the more resilient the genotype.

References

Ghazvini, H., Pour-Aboughadareh, A., Jasemi, S.S. et al. A Framework for Selection of High-Yielding and Drought-tolerant Genotypes of Barley: Applying Yield-Based Indices and Multi-index Selection Models. Journal of Crop Health 76, 601–616 (2024). https://doi.org/10.1007/s10343-024-00981-1

Author

Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva

Examples

# \donttest{
library(Breeding)

data("aveia")
with(aveia,estresse(GEN,MC,MG,index = "ALL",bygen=T))
#> Loading required package: ggplot2
#> Loading required package: tidyr
#> Loading required package: viridis
#> Loading required package: viridisLite
#> 
#> ---------------------------------------------------------------------------------------------
#> Índices de Estresse
#> ---------------------------------------------------------------------------------------------
#>    GEN       STI        YI      GMP        MP        MH       SSI       YSI
#> 1   G1 0.6843575 0.6843575 1.233995 0.7613715 0.3747927 1.1866617 0.6843575
#> 2  G10 0.7628362 0.7628362 1.488428 1.1077083 0.2690707 0.7998926 0.7628362
#> 3  G11 0.7172237 0.7172237 1.372668 0.9421094 0.3293413 1.0143855 0.7172237
#> 4  G12 0.7589286 0.7589286 1.626175 1.3222222 0.2741117 0.8172584 0.7589286
#> 5  G13 0.6997519 0.6997519 1.404643 0.9865104 0.3532847 1.1039536 0.6997519
#> 6  G14 0.6722222 0.6722222 1.229837 0.7562500 0.3920266 1.2545304 0.6722222
#> 7   G2 0.7753623 0.7753623 1.518943 1.1535938 0.2530612 0.7454053 0.7753623
#> 8   G3 0.7288136 0.7288136 1.259216 0.7928125 0.3137255 0.9573397 0.7288136
#> 9   G4 0.7648579 0.7648579 1.410230 0.9943750 0.2664714 0.7909777 0.7648579
#> 10  G5 0.7218543 0.7218543 1.425474 1.0159877 0.3230769 0.9913736 0.7218543
#> 11  G6 0.6934673 0.6934673 1.380972 0.9535417 0.3620178 1.1372745 0.6934673
#> 12  G7 0.7043478 0.7043478 1.608571 1.2937500 0.3469388 1.0799619 0.7043478
#> 13  G8 0.6726058 0.6726058 1.534318 1.1770660 0.3914780 1.2523478 0.6726058
#> 14  G9 0.7193878 0.7193878 1.385340 0.9595833 0.3264095 1.0035942 0.7193878
#>          RSI
#> 1  1.0522445
#> 2  0.9439922
#> 3  1.0040263
#> 4  0.9488527
#> 5  1.0290954
#> 6  1.0712401
#> 7  0.9287418
#> 8  0.9880599
#> 9  0.9414970
#> 10 0.9975856
#> 11 1.0384216
#> 12 1.0223804
#> 13 1.0706293
#> 14 1.0010060

# }