Heterosis and Heterobeltiosis

Calculation of heterosis and heterobeltiosis parameters of hybrids

heterose(GEN, GM, GP, PR, REP, param = "all")

Arguments

GEN

The column with the genotype name

GM

The column with the average of the maternal parent

GP

The column with the average of the paternal parent

PR

The column with the average of the progeny

REP

The column with the repetitions

param

Value to determine the parameter to be calculated. Default is “all”. To calculate heterosis only, use “het”. To calculate only heterobeltiosis, use “hetb”.

Author

Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva

Examples

# \donttest{
library(Breeding)

data("maize")
#Extract heterosis and heterobeltiosis
with(maize,heterose(HIB,GM,GP,P,stat="all"))
#> Error in heterose(HIB, GM, GP, P, stat = "all"): unused argument (stat = "all")

#Only extract heterosis
with(maize,heterose(HIB,GM,GP,P,param = "het"))
#> Error in eval(substitute(expr), data, enclos = parent.frame()): object 'HIB' not found

#Extract only heterobeltiosis
with(maize,heterose(HIB,GM,GP,P,param = "hetb"))
#> Error in eval(substitute(expr), data, enclos = parent.frame()): object 'HIB' not found
# }