Ear Indexes

Estimating the viability index from the combination of two field variables.

indviab(
  GEN,
  var1,
  var2,
  ylab = "Index",
  xlab = "Genotype",
  stat = "all",
  plot = F
)

Arguments

var1

The column containing the first variable

var2

The column containing the second variable

genot

The column with the name of the genotypes

References

Rigotti, E. J., Carvalho, I. R., Loro, M. V., Pradebon, L. C., Dalla Roza, J. P., & Sangiovo, J. P. (2024). Seed and grain yield and quality of wheat subjected to advanced harvest using a physiological ripening process. Revista Engenharia na Agricultura - REVENG, 32(Contínua), 54–64. https://doi.org/10.13083/reveng.v32i1.17394

Author

Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva

Examples

# \donttest{
library(EstimateBreed)

data("trigo")
#Ear viability index
with(trigo,indviab(TEST,NGE,NEE))
#> 
#> -----------------------------------------------------------------
#> Índice de Viabilidade Geral
#> -----------------------------------------------------------------
#>    Genótipo    Índice
#> 1        T1 0.7270983
#> 2        T2 0.7927928
#> 3        T3 0.7572954
#> 4        T4 0.7951413
#> 5        T5 0.7578163
#> 6        T6 0.7882816
#> 7        T7 0.7781609
#> 8        T8 0.7793054
#> 9        T9 0.8141160
#> 10      T10 0.8498195
#> 11      T11 0.8438819
#> 12      T12 0.8014337
#> 13      T13 0.7719665
#> 14      T14 0.8245132
#> 15      T15 0.8051075
#> 16      T16 0.7022956
#> 17      T17 0.7624164
#> 18      T18 0.7773723
#> 19      T19 0.7438854

#Ear harvest index
with(trigo,indviab(TEST,MGE,ME))
#> 
#> -----------------------------------------------------------------
#> Índice de Viabilidade Geral
#> -----------------------------------------------------------------
#>    Genótipo    Índice
#> 1        T1 0.7530794
#> 2        T2 0.7211403
#> 3        T3 0.7250430
#> 4        T4 0.7276371
#> 5        T5 0.7527310
#> 6        T6 0.7404977
#> 7        T7 0.7513300
#> 8        T8 0.7555921
#> 9        T9 0.7464156
#> 10      T10 0.7605856
#> 11      T11 0.7486986
#> 12      T12 0.7393468
#> 13      T13 0.7484841
#> 14      T14 0.7646920
#> 15      T15 0.7515557
#> 16      T16 0.6933280
#> 17      T17 0.7543860
#> 18      T18 0.7340161
#> 19      T19 0.7276199

#Spikelet deposition index in the ear
#'with(trigo,indviab(TEST,NEE,CE))
# }