Industrial quality of wheat

Function for determining industrial quality indices of wheat genotypes, described by Szareski et al. (2019).

is_qindustrial(GEN, NQ, W, PTN)

Arguments

GEN

The column with the genotype name

NQ

The column with the falling number

W

The column with the gluten force (W)

PTN

The column with the protein values

References

Szareski, V. J., Carvalho, I. R., Kehl, K., Levien, A. M., Lautenchleger, F., Barbosa, M. H., ... & Aumonde, T. Z. (2019). Genetic and phenotypic multi-character approach applied to multivariate models for wheat industrial quality analysis. Genetics and Molecular Research, 18(3), 1-14.

Author

Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva

Examples

# \donttest{
library(EstimateBreed)

data("ptn")
with(ptn,is_qindustrial(Cult,NQ,W,PTN))
#> 
#> -----------------------------
#> Índice de Qualidade do Trigo
#> -----------------------------
#>     GEN     Index
#> 1     1  912.4249
#> 2     1  892.5425
#> 3     1  918.5076
#> 4     1  770.5968
#> 5     1  998.0063
#> 6     1  875.6489
#> 7     1  792.9283
#> 8     1  795.5081
#> 9     1  844.1790
#> 10    1  809.0006
#> 11    1  753.3499
#> 12    1  852.3557
#> 13    1  684.9554
#> 14    1  856.7863
#> 15    1  679.4566
#> 16    1  744.1493
#> 17    1  589.6304
#> 18    1  680.1454
#> 19    1  653.1867
#> 20    1  747.7734
#> 21    1  849.3140
#> 22    1  865.6657
#> 23    1  873.6390
#> 24    1  806.9315
#> 25    1  826.2587
#> 26    1  955.7345
#> 27    1  731.2019
#> 28    1  670.4775
#> 29    1  798.0012
#> 30    1  702.2320
#> 31    1  843.3594
#> 32    1 1064.0202
#> 33    1  733.4408
#> 34    1  853.2604
#> 35    1  898.3795
#> 36    1  759.8502
#> 37    1  737.9059
#> 38    1  692.9529
#> 39    1  767.2468
#> 40    1  774.6997
#> 41    1  773.0770
#> 42    1  785.3395
#> 43    1  745.2220
#> 44    1  732.0998
#> 45    1  949.6631
#> 46    1  882.4102
#> 47    1  872.8781
#> 48    1  856.8044
#> 49    1  853.5677
#> 50    1  983.0854
#> 51    1  735.4183
#> 52    1  862.9514
#> 53    1  880.8424
#> 54    1  704.3688
#> 55    1  731.2132
#> 56    1  692.3060
#> 57    1  675.1677
#> 58    1  614.6270
#> 59    1  611.7484
#> 60    1  797.0146
#> 61    1 1091.5006
#> 62    1 1116.6560
#> 63    1 1025.6324
#> 64    1  914.4803
#> 65    1  898.8017
#> 66    1 1083.7927
#> 67    1 1074.3478
#> 68    1  897.2667
#> 69    1  888.9587
#> 70    1 1008.2798
#> 71    1  948.5963
#> 72    1  958.4957
#> 73    1  959.5408
#> 74    1  916.6221
#> 75    1 1116.1386
#> 76    1  870.3092
#> 77    1  795.3855
#> 78    1  928.0825
#> 79    1  837.7273
#> 80    1  942.7164
#> 81    1  970.9795
#> 82    1  773.1692
#> 83    1 1055.7015
#> 84    1 1007.2160
#> 85    1  982.4979
#> 86    1 1150.8402
#> 87    1  974.7094
#> 88    1 1093.1799
#> 89    1 1029.8426
#> 90    1  934.3712
#> 91    1 1049.8283
#> 92    1  975.1218
#> 93    1 1012.2489
#> 94    1  830.8847
#> 95    1  874.6840
#> 96    1 1027.1142
#> 97    1  842.9525
#> 98    1  797.7181
#> 99    1  752.9434
#> 100   1  801.0223
#> 101   1  858.4372
#> 102   1  796.6222
#> 103   1  852.5340
#> 104   1  939.2169
#> 105   1 1241.5832
#> 106   1 1128.9604
#> 107   1 1144.4751
#> 108   1 1037.4963
#> 109   1  991.5541
#> 110   1 1076.2188
#> 111   1  987.9884
#> 112   1  919.9046
#> 113   1 1019.9302
#> 114   1  974.1786
#> 115   1 1059.9791
#> 116   1  884.4041
#> 117   1  808.6763
#> 118   1  705.9153
#> 119   1  868.1209
#> 120   1  666.5786
#> 121   2  702.2749
#> 122   2  594.7805
#> 123   2  607.7539
#> 124   2  565.8675
#> 125   2  904.7164
#> 126   2  866.0040
#> 127   2  808.2424
#> 128   2  797.8514
#> 129   2  731.6104
#> 130   2  800.8481
#> 131   2  692.6253
#> 132   2  870.5668
#> 133   2  675.7356
#> 134   2  662.1105
#> 135   2  666.7259
#> 136   2  839.9460
#> 137   2  610.1384
#> 138   2  613.0998
#> 139   2  610.8756
#> 140   2  686.9016
#> 141   2  794.1253
#> 142   2  617.1079
#> 143   2  816.9343
#> 144   2  802.1062
#> 145   2  738.9502
#> 146   2  838.0451
#> 147   2  713.1989
#> 148   2  718.9239
#> 149   2  862.2106
#> 150   2  770.2778
#> 151   2  823.6374
#> 152   2  777.5921
#> 153   2  825.1436
#> 154   2  725.4914
#> 155   2  779.2081
#> 156   2  689.0366
#> 157   2  541.1602
#> 158   2  737.5884
#> 159   2  575.2496
#> 160   2  562.8253
#> 161   2  790.9559
#> 162   2  778.0791
#> 163   2  864.8169
#> 164   2  740.2675
#> 165   2  922.2931
#> 166   2  747.3526
#> 167   2  875.0241
#> 168   2  917.2467
#> 169   2  672.1424
#> 170   2  786.4042
#> 171   2  734.0766
#> 172   2  812.8999
#> 173   2  613.5542
#> 174   2  639.3775
#> 175   2  580.7318
#> 176   2  570.2818
#> 177   2  708.6492
#> 178   2  786.4640
#> 179   2  725.7299
#> 180   2  627.6354
#> 181   2  447.0015
#> 182   2  656.4616
#> 183   2  586.7692
#> 184   2  702.2810
#> 185   2  643.3714
#> 186   2  631.7590
#> 187   2  676.4520
#> 188   2  680.9943
#> 189   2  604.2840
#> 190   2  597.8418
#> 191   2  632.0996
#> 192   2  610.8676
#> 193   2  531.8099
#> 194   2  508.1308
#> 195   2  466.2893
#> 196   2  485.2009
#> 197   2  572.5793
#> 198   2  511.6029
#> 199   2  479.6929
#> 200   2  615.5556
#> 201   2  493.5721
#> 202   2  543.3224
#> 203   2  502.3631
#> 204   2  499.8685
#> 205   2  477.9205
#> 206   2  524.5346
#> 207   2  659.0171
#> 208   2  665.3472
#> 209   2  585.7079
#> 210   2  518.3638
#> 211   2  572.9593
#> 212   2  488.8222
#> 213   2  435.2350
#> 214   2  441.4334
#> 215   2  396.5963
#> 216   2  498.6340
#> 217   2  444.2175
#> 218   2  512.3823
#> 219   2  412.0274
#> 220   2  531.0281
#> 221   2  442.0211
#> 222   2  422.1316
#> 223   2  407.9313
#> 224   2  512.9456
#> 225   2  418.4802
#> 226   2  462.7388
#> 227   2  562.9811
#> 228   2  577.4690
#> 229   2  456.7527
#> 230   2  528.9312
#> 231   2  502.1023
#> 232   2  449.0746
#> 233   2  439.8884
#> 234   2  482.4520
#> 235   2  490.3901
#> 236   2  464.2164
#> 237   2  612.1015
#> 238   2  534.9005
#> 239   2  461.5797
#> 240   2  485.2302
#> 241   3  964.4989
#> 242   3 1032.5531
#> 243   3 1013.3826
#> 244   3  981.8408
#> 245   3 1021.8254
#> 246   3 1069.5844
#> 247   3 1032.6390
#> 248   3  993.4557
#> 249   3  847.2232
#> 250   3  831.3734
#> 251   3  919.8266
#> 252   3  972.8833
#> 253   3  838.2012
#> 254   3 1069.7394
#> 255   3  925.2339
#> 256   3  835.3730
#> 257   3  700.3431
#> 258   3  709.1868
#> 259   3  942.7747
#> 260   3  774.9740
#> 261   3  757.6954
#> 262   3  778.0325
#> 263   3  710.1674
#> 264   3  849.3244
#> 265   3  858.5389
#> 266   3 1020.6546
#> 267   3  970.9244
#> 268   3  928.7547
#> 269   3  811.1050
#> 270   3  747.6986
#> 271   3  896.1987
#> 272   3  673.8037
#> 273   3  718.8163
#> 274   3  893.1184
#> 275   3  768.4132
#> 276   3  731.4784
#> 277   3  733.0556
#> 278   3  671.2705
#> 279   3  635.7211
#> 280   3  538.0404
#> 281   3  619.1492
#> 282   3  600.5865
#> 283   3  627.7330
#> 284   3  755.2195
#> 285   3  610.8615
#> 286   3  729.1626
#> 287   3  796.2880
#> 288   3  763.7722
#> 289   3  573.3162
#> 290   3  585.9248
#> 291   3  604.3091
#> 292   3  581.5374
#> 293   3  381.0993
#> 294   3  387.2909
#> 295   3  471.6837
#> 296   3  426.8832
#> 297   3  524.6091
#> 298   3  601.7776
#> 299   3  634.1819
#> 300   3  515.2805
#> 301   3  532.0387
#> 302   3  531.5593
#> 303   3  490.9973
#> 304   3  509.0820
#> 305   3  632.2536
#> 306   3  517.0077
#> 307   3  693.7124
#> 308   3  572.6731
#> 309   3  448.4344
#> 310   3  474.1162
#> 311   3  475.3466
#> 312   3  500.7921
#> 313   3  478.5743
#> 314   3  519.0639
#> 315   3  392.4081
#> 316   3  373.0645
#> 317   3  507.6096
#> 318   3  519.0723
#> 319   3  536.6543
#> 320   3  603.5978
#> 321   3  549.9354
#> 322   3  466.7542
#> 323   3  448.6984
#> 324   3  442.5235
#> 325   3  479.2126
#> 326   3  489.8045
#> 327   3  469.2421
#> 328   3  565.8996
#> 329   3  445.2533
#> 330   3  423.8381
#> 331   3  423.6454
#> 332   3  391.0568
#> 333   3  471.3225
#> 334   3  318.6794
#> 335   3  398.4273
#> 336   3  304.6876
#> 337   3  444.2505
#> 338   3  388.2885
#> 339   3  352.3697
#> 340   3  384.5161
#> 341   3  854.9956
#> 342   3  858.4047
#> 343   3  935.2091
#> 344   3  805.1317
#> 345   3  729.7046
#> 346   3  961.1425
#> 347   3  879.4640
#> 348   3  899.6989
#> 349   3 1063.2673
#> 350   3  939.4229
#> 351   3  998.5939
#> 352   3  935.6744
#> 353   3  774.9522
#> 354   3  821.3793
#> 355   3  804.2292
#> 356   3  970.3711
#> 357   3  837.0574
#> 358   3  935.8704
#> 359   3  721.8175
#> 360   3  829.1253
#> 
#> -----------------------------
#> Desvios
#> -----------------------------
#>      sd_FN     sd_W   sd_PTN
#> 1 24.59698 44.65365 1.993301
# }