ISGR - Genetic Selection Index for Resilience

Estimation of the selection index for environmental resilience (Bandeira et al., 2024).

isgr(GEN, ENV, NG, MG, CICLO, req = 3.5, stage = NULL)

Arguments

GEN

Column referring to genotypes. Lines must have the prefix “L” before the number. Ex: L139.

ENV

The column for the selection environment.

NG

Number of grains of all genotypes evaluated

MG

Grain mass of all genotypes evaluated

CICLO

Number of days in the cycle to define rainfall ideal (value of 3.5 mm per day). Can be changed manually in the 'req' argument.

req

Average daily water demand for the soybean crop (standard 3.5 mm). May change depending on the phenological stage.

stage

Parameter to define the phenological stage the crop is in Use “veg” for vegetative and “rep” for reproductive, if the evaluations have only been carried out in a given period.

References

Bandeira, W. J. A., Carvalho, I. R., Loro, M. V., da Silva, J. A. G., Dalla Roza, J. P., Scarton, V. D. B., Bruinsma, G. M. W., & Pradebon, L. C. (2024). Identifying soybean progenies with high grain productivity and stress resilience to abiotic stresses. Aust J Crop Sci, 18(12), 825-830. https://doi.org/10.21475/ajcs.24.18.12.p98

Author

Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva

Examples

# \donttest{
library(EstimateBreed)
#Obtain environmental deviations
data("desvamb")
with(desvamb, desv_clim(ENV,TMED,PREC))
#> # A tibble: 3 × 5
#>   ENV   STMED TMEDR SPREC PRECIR
#>   <chr> <dbl> <dbl> <dbl>  <dbl>
#> 1 E1     2.65  24.8  5.46   339.
#> 2 E2     3.65  23.8  5.27   344.
#> 3 E3     2.81  24.5  5.47   362.

#Calculate the ISGR
data("genot")
with(genot, isgr(GEN,ENV,NG,MG,CICLO))
#>     Gen Env      ISGR
#> 26 L454  E1  6.489941
#> 22 L455  E1  7.084315
#> 19 L541  E1  7.653157
#> 18 L367  E1  7.862185
#> 16 L380  E1  8.329434
#> 12 L393  E1  9.638909
#> 10 L439  E1 10.552056
#> 28 L298  E3 12.209433
#> 30 L358  E2 23.347984
#> 29 L346  E2 23.793351
#> 27 L195  E2 24.719927
#> 25 L179  E2 25.747317
#> 24 L359  E2 26.300686
#> 23 L345  E2 26.886419
#> 1  L445  E1 27.255375
#> 21 L185  E2 28.211433
#> 20 L310  E2 28.942165
#> 17 L178  E2 31.418785
#> 15 L261  E2 33.424611
#> 14 L269  E2 34.605133
#> 13 L209  E2 35.959423
#> 11 L263  E2 39.127798
#> 9  L201  E2 43.145922
#> 8  L299  E2 45.686042
#> 7  L152  E2 48.926278
#> 6   L26  E2 52.988109
#> 5  L166  E2 57.596139
#> 4  L155  E2 64.251152
#> 3  L277  E2 74.756384
#> 2  L162  E2 86.543916

#Define the water requirement per stage
with(genot, isgr(GEN,ENV,NG,MG,CICLO,req=5,stage="rep"))
#> Error in isgr(GEN, ENV, NG, MG, CICLO, req = 5, stage = "rep"): unused argument (stage = "rep")
# }