Hectolitre weight of cereals

Useful function for characterizing the hectolitre weight (HW) of experiments with cereals.

ph(GEN, PESO, crop = "trigo", stat = "all")

Arguments

GEN

The column with the genotype name

PESO

Weight obtained on a 1qt lt scale, as determined by the Rules for Seed Analysis (RAS), Ministry of Agriculture, Livestock and Supply (2009).

Value

Returns the estimated value of the hectolitre weight (HW) for the ceral selecionado.

References

Brasil. Ministério da Agricultura, Pecuária e Abastecimento. Secretaria de Defesa Agropecuária. Regras para Análise de Sementes. Brasília: MAPA/ACS, 2009. 399 p. ISBN 978-85-99851-70-8.

Author

Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva

Examples

# \donttest{
library(EstimateBreed)

GEN <- rep(paste("G", 1:5, sep=""), each = 3)
REP <- rep(1:3, times = 5)
MG <- c(78.5, 80.2, 79.1, 81.3, 82.0, 80.8, 76.9, 78.1, 77.5, 83.2,
84.1, 82.9, 77.4, 78.9, 79.3)

data <- data.frame(GEN, REP, MG)

with(data,ph(GEN,MG,crop="trigo"))
#>    GEN PESO       PH
#> 1   G1 78.5 25.53219
#> 2   G1 80.2 26.30029
#> 3   G1 79.1 25.80328
#> 4   G2 81.3 26.79729
#> 5   G2 82.0 27.11356
#> 6   G2 80.8 26.57138
#> 7   G3 76.9 24.80928
#> 8   G3 78.1 25.35146
#> 9   G3 77.5 25.08037
#> 10  G4 83.2 27.65575
#> 11  G4 84.1 28.06239
#> 12  G4 82.9 27.52020
#> 13  G5 77.4 25.03519
#> 14  G5 78.9 25.71292
#> 15  G5 79.3 25.89365

#Extract the average PH per genotype
with(data,ph(GEN,MG,crop="trigo",stat="mean"))
#> Error in ph(GEN, MG, crop = "trigo", stat = "mean"): unused argument (stat = "mean")
# }