Peeling Index and Industrial Yield

Calculating the Hulling Index and Industrial Yield of White Oats

rend_ind(GEN, NG2M, MG, MC, RG, stat = "all", ...)

Arguments

GEN

The column with the name of the genotypes.

NG2M

The column with values for the number of grains larger than 2mm.

MG

The column with grain mass values.

MC

The column with karyopsis mass values.

RG

The column with the grain yield values (kg per ha).

stat

Logical argument. Use “all” to keep all the observations or “mean” to extract the overall average.

Author

Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva

Examples

# \donttest{
library(EstimateBreed)

data("aveia")
# Calculate the industrial yield without extracting the average
with(aveia, rend_ind(GEN,NG2M,MG,MC,RG))
#> 
#> -----------------------------------------------------------------
#> Índice de Descasque e Rendimento Industrial
#> -----------------------------------------------------------------
#>    GEN        ID        RI
#> 1   G1 0.6421053  248.0992
#> 2   G1 0.7058824  259.2127
#> 3   G1 0.7263158  401.2875
#> 4   G1 0.6626506  142.5706
#> 5   G2 0.7636364  513.0272
#> 6   G2 0.7589286  473.7050
#> 7   G2 0.7882353  421.0346
#> 8   G2 0.7943925  735.9787
#> 9   G3 0.7659574  517.4221
#> 10  G3 0.7333333  198.2880
#> 11  G3 0.6808511  189.6073
#> 12  G3 0.7362637  257.6884
#> 13  G4 0.7844828 1309.5622
#> 14  G4 0.7551020  557.8128
#> 15  G4 0.7608696  608.3568
#> 16  G4 0.7530864  299.6963
#> 17  G5 0.7346939  698.7208
#> 18  G5 0.7583333  687.4180
#> 19  G5 0.6547619  144.2462
#> 20  G6 0.7083333  421.3280
#> 21  G6 0.7547170  376.3683
#> 22  G6 0.6822430  201.4227
#> 23  G6 0.6179775  241.5106
#> 24  G7 0.6893939  418.2682
#> 25  G7 0.7475728  409.8992
#> 26  G7 0.6818182  258.0423
#> 27  G8 0.6694915  779.4310
#> 28  G8 0.7117117  319.0226
#> 29  G8 0.6400000  436.5363
#> 30  G8 0.6666667  326.3160
#> 31  G9 0.7019231  679.3385
#> 32  G9 0.8080808  482.9684
#> 33  G9 0.7809524  685.0514
#> 34  G9 0.5595238  159.5463
#> 35 G10 0.7570093  920.7979
#> 36 G10 0.7578947  388.1068
#> 37 G10 0.7767857  617.4250
#> 38 G10 0.7578947  579.7885
#> 39 G11 0.7589286  536.0166
#> 40 G11 0.7943925  600.1238
#> 41 G11 0.7282609  281.4175
#> 42 G11 0.5384615  271.5605
#> 43 G12 0.7678571  694.0707
#> 44 G12 0.7631579  900.9143
#> 45 G12 0.7572816  665.8663
#> 46 G12 0.7478992  957.2217
#> 47 G13 0.6976744  503.7428
#> 48 G13 0.5543478  265.6668
#> 49 G13 0.7565217  575.4152
#> 50 G13 0.7636364  585.0671
#> 51 G14 0.6547619  374.1069
#> 52 G14 0.7325581  662.9329
#> 53 G14 0.5609756  330.6989
#> 54 G14 0.7222222  575.8133

# Calculate the industrial yield by extracting the average per genotype
with(aveia, rend_ind(GEN,NG2M,MG,MC,RG,stat="mean"))
#> 
#> -----------------------------------------------------------------
#> Índice de Descasque e Rendimento Industrial (Média por Genótipo)
#> -----------------------------------------------------------------
#>    GEN        ID       RI
#> 1   G1 0.6842385 262.7925
#> 2  G10 0.7623961 626.5295
#> 3  G11 0.7050109 422.2796
#> 4  G12 0.7590489 804.5182
#> 5  G13 0.6930451 482.4730
#> 6  G14 0.6676295 485.8880
#> 7   G2 0.7762982 535.9364
#> 8   G3 0.7291014 290.7514
#> 9   G4 0.7633852 693.8570
#> 10  G5 0.7159297 510.1284
#> 11  G6 0.6908177 310.1574
#> 12  G7 0.7062616 362.0699
#> 13  G8 0.6719675 465.3265
#> 14  G9 0.7126200 501.7261
# }