Risk of Disease Occurrence in Soybeans

Calculation of the Risk of Disease Occurrence in Soybeans as a Function of Variables meteorological variables (Engers et al., 2024).

risco(DIA, MES, TEMP, UR, doença = "rust", plot = F)

Arguments

DIA

The column for the day of the month.

MES

The column for the month of the year (numeric value).

TEMP

A coluna da temperatura média do ar (em ºC).

UR

The relative humidity column (in %).

doença

Define the soybean disease (Standard = “rust”).

plot

Plot a graph of the accumulation (Default is F (FALSE)).

References

de Oliveira Engers, L.B., Radons, S.Z., Henck, A.U. et al. Evaluation of a forecasting system to facilitate decision-making for the chemical control of Asian soybean rust. Trop. plant pathol. 49, 539–546 (2024). https://doi.org/10.1007/s40858-024-00649-1

Author

Willyan Júnior Adorian Bandeira

Ivan Ricardo Carvalho

Murilo Vieira Loro

Leonardo Cesar Pradebon

José Antonio Gonzalez da Silva

Examples

# \donttest{
library(Breeding)

# Rust Risk Prediction
data("clima")
with(clima, risco(DY, MO, TMED, RH, doença = "ferrugem"))
#>     DIA MES UR
#> 1     1   1  8
#> 2     2   1 13
#> 3     3   1 15
#> 4     4   1 16
#> 5     5   1 15
#> 6     6   1  8
#> 7     7   1 13
#> 8     8   1 14
#> 9     9   1 13
#> 10   10   1 13
#> 11   11   1 12
#> 12   12   1 15
#> 13   13   1 14
#> 14   14   1 11
#> 15   15   1 10
#> 16   16   1  8
#> 17   17   1 12
#> 18   18   1 13
#> 19   19   1 14
#> 20   20   1 14
#> 21   21   1  7
#> 22   22   1  3
#> 23   23   1 10
#> 24   24   1  6
#> 25   25   1 11
#> 26   26   1 14
#> 27   27   1 10
#> 28   28   1 15
#> 29   29   1 18
#> 30   30   1 16
#> 31   31   1 15
#> 32    1   2 17
#> 33    2   2 16
#> 34    3   2 15
#> 35    4   2 15
#> 36    5   2 15
#> 37    6   2 13
#> 38    7   2 16
#> 39    8   2 14
#> 40    9   2  3
#> 41   10   2 14
#> 42   11   2 10
#> 43   12   2 11
#> 44   13   2 13
#> 45   14   2 13
#> 46   15   2 13
#> 47   16   2 13
#> 48   17   2 12
#> 49   18   2 13
#> 50   19   2 14
#> 51   20   2 15
#> 52   21   2 13
#> 53   22   2 12
#> 54   23   2 13
#> 55   24   2  8
#> 56   25   2 12
#> 57   26   2 14
#> 58   27   2 13
#> 59   28   2  7
#> 60    1   3  6
#> 61    2   3 13
#> 62    3   3 14
#> 63    4   3 13
#> 64    5   3 11
#> 65    6   3 13
#> 66    7   3 12
#> 67    8   3 11
#> 68    9   3  7
#> 69   10   3  6
#> 70   11   3 11
#> 71   12   3 15
#> 72   13   3 15
#> 73   14   3 14
#> 74   15   3 16
#> 75   16   3 14
#> 76   17   3 15
#> 77   18   3  9
#> 78   19   3 14
#> 79   20   3 17
#> 80   21   3 16
#> 81   22   3 15
#> 82   23   3 13
#> 83   24   3 10
#> 84   25   3 16
#> 85   26   3 13
#> 86   27   3 16
#> 87   28   3 15
#> 88   29   3 14
#> 89   30   3 11
#> 90   31   3 13
#> 91    1   4 16
#> 92    2   4 15
#> 93    3   4 14
#> 94    4   4 12
#> 95    5   4 12
#> 96    6   4 13
#> 97    7   4  8
#> 98    8   4 10
#> 99    9   4 10
#> 100  10   4  8
#> 101  11   4 12
#> 102  12   4  8
#> 103  13   4 11
#> 104  14   4 15
#> 105  15   4 12
#> 106  16   4  7
#> 107  17   4  3
#> 108  18   4  6
#> 109  19   4  5
#> 110  20   4  4
#> 111  21   4  5
#> 112  22   4  0
#> 113  23   4 11
#> 114  24   4  9
#> 115  25   4  7
#> 116  26   4  7
#> 117  27   4  6
#> 118  28   4  6
#> 119  29   4  0
#> 120  30   4  2
#> 121   1   5  0
#> 122   2   5  0
#> 123   3   5  2
#> 124   4   5  7
#> 125   5   5  5
#> 126   6   5  4
#> 127   7   5  0
#> 128   8   5  0
#> 129   9   5  0
#> 130  10   5  0
#> 131  11   5  0
#> 132  12   5  6
#> 133  13   5  5
#> 134  14   5  5
#> 135  15   5  7
#> 136  16   5  8
#> 137  17   5  1
#> 138  18   5  3
#> 139  19   5  9
#> 140  20   5  7
#> 141  21   5  2
#> 142  22   5  0
#> 143  23   5  0
#> 144  24   5  0
#> 145  25   5  7
#> 146  26   5  3
#> 147  27   5  4
#> 148  28   5  4
#> 149  29   5  3
#> 150  30   5  0
#> 151  31   5  3
#> 152   1   6  6
#> 153   2   6  7
#> 154   3   6  7
#> 155   4   6  7
#> 156   5   6  6
#> 157   6   6  3
#> 158   7   6  0
#> 159   8   6  0
#> 160   9   6 12
#> 161  10   6 17
#> 162  11   6 12
#> 163  12   6  0
#> 164  13   6  0
#> 165  14   6  5
#> 166  15   6  7
#> 167  16   6  6
#> 168  17   6  9
#> 169  18   6 16
#> 170  19   6  8
#> 171  20   6  7
#> 172  21   6  0
#> 173  22   6  3
#> 174  23   6  0
#> 175  24   6  0
#> 176  25   6  0
#> 177  26   6  1
#> 178  27   6  8
#> 179  28   6  7
#> 180  29   6  0
#> 181  30   6  1
#> 182   1   7  0
#> 183   2   7  0
#> 184   3   7  0
#> 185   4   7  0
#> 186   5   7  0
#> 187   6   7  0
#> 188   7   7  0
#> 189   8   7  0
#> 190   9   7  0
#> 191  10   7  0
#> 192  11   7  0
#> 193  12   7  0
#> 194  13   7  6
#> 195  14   7  0
#> 196  15   7  0
#> 197  16   7  0
#> 198  17   7  0
#> 199  18   7  5
#> 200  19   7  0
#> 201  20   7  0
#> 202  21   7  0
#> 203  22   7  0
#> 204  23   7  0
#> 205  24   7  0
#> 206  25   7  0
#> 207  26   7  0
#> 208  27   7  0
#> 209  28   7  0
#> 210  29   7  0
#> 211  30   7  1
#> 212  31   7  0
#> 213   1   8  0
#> 214   2   8  0
#> 215   3   8  0
#> 216   4   8  0
#> 217   5   8  0
#> 218   6   8  0
#> 219   7   8  0
#> 220   8   8  0
#> 221   9   8  6
#> 222  10   8  6
#> 223  11   8  6
#> 224  12   8  5
#> 225  13   8  0
#> 226  14   8  0
#> 227  15   8  0
#> 228  16   8  0
#> 229  17   8  0
#> 230  18   8  6
#> 231  19   8  7
#> 232  20   8  1
#> 233  21   8  7
#> 234  22   8  1
#> 235  23   8  0
#> 236  24   8  0
#> 237  25   8  0
#> 238  26   8  0
#> 239  27   8  0
#> 240  28   8  0
#> 241  29   8  3
#> 242  30   8  0
#> 243  31   8  0
#> 244   1   9  0
#> 245   2   9  0
#> 246   3   9  0
#> 247   4   9  0
#> 248   5   9  0
#> 249   6   9  0
#> 250   7   9  0
#> 251   8   9  0
#> 252   9   9  0
#> 253  10   9  0
#> 254  11   9  0
#> 255  12   9  0
#> 256  13   9  0
#> 257  14   9  1
#> 258  15   9  7
#> 259  16   9  0
#> 260  17   9  0
#> 261  18   9  0
#> 262  19   9  0
#> 263  20   9  0
#> 264  21   9  0
#> 265  22   9  5
#> 266  23   9  3
#> 267  24   9  0
#> 268  25   9  0
#> 269  26   9  4
#> 270  27   9 11
#> 271  28   9 12
#> 272  29   9 13
#> 273  30   9  9
#> 274   1  10  8
#> 275   2  10  8
#> 276   3  10  6
#> 277   4  10  3
#> 278   5  10  6
#> 279   6  10  8
#> 280   7  10  8
#> 281   8  10  7
#> 282   9  10  5
#> 283  10  10 15
#> 284  11  10  8
#> 285  12  10 14
#> 286  13  10  7
#> 287  14  10  4
#> 288  15  10  2
#> 289  16  10  0
#> 290  17  10  0
#> 291  18  10  0
#> 292  19  10  3
#> 293  20  10  1
#> 294  21  10 13
#> 295  22  10 12
#> 296  23  10  6
#> 297  24  10  2
#> 298  25  10  1
#> 299  26  10  0
#> 300  27  10  0
#> 301  28  10  6
#> 302  29  10 13
#> 303  30  10  8
#> 304  31  10 12
#> 305   1  11  7
#> 306   2  11  7
#> 307   3  11  6
#> 308   4  11  3
#> 309   5  11  0
#> 310   6  11  0
#> 311   7  11  0
#> 312   8  11  0
#> 313   9  11  0
#> 314  10  11  0
#> 315  11  11  6
#> 316  12  11  6
#> 317  13  11  6
#> 318  14  11 10
#> 319  15  11  7
#> 320  16  11  0
#> 321  17  11  0
#> 322  18  11  0
#> 323  19  11  0
#> 324  20  11  7
#> 325  21  11  4
#> 326  22  11  3
#> 327  23  11 10
#> 328  24  11  6
#> 329  25  11  2
#> 330  26  11  0
#> 331  27  11  0
#> 332  28  11  0
#> 333  29  11  0
#> 334  30  11  0
#> 335   1  12  2
#> 336   2  12  8
#> 337   3  12 13
#> 338   4  12 15
#> 339   5  12 12
#> 340   6  12 12
#> 341   7  12  7
#> 342   8  12 12
#> 343   9  12 13
#> 344  10  12 13
#> 345  11  12 13
#> 346  12  12 15
#> 347  13  12 14
#> 348  14  12 13
#> 349  15  12 14
#> 350  16  12 11
#> 351  17  12 11
#> 352  18  12 12
#> 353  19  12 14
#> 354  20  12 13
#> 355  21  12 11
#> 356  22  12  8
#> 357  23  12  7
#> 358  24  12 11
#> 359  25  12 15
#> 360  26  12 14
#> 361  27  12 13
#> 362  28  12 13
#> 363  29  12 13
#> 364  30  12 13
#> 365  31  12 13
#> 
#> --------------------------------------------------
#> Risco de Ocorrência de Ferrugem Asiática
#> --------------------------------------------------
#>    Mês   riscoUR riscoTEMP riscoTotal     riscoRel
#> 1    1 12.706058  26.83114  340.91801  77.28258522
#> 2   10  2.132149  24.50724   52.25308   9.70383628
#> 3   11  1.010342  22.68888   22.92353   2.83755630
#> 4   12  5.045327  27.28285  137.65089  29.69614071
#> 5    2 13.377456  26.68142  356.92951  81.03100483
#> 6    3 16.307827  26.85558  437.95622 100.00000000
#> 7    4  4.674734  21.95465  102.63217  21.49798046
#> 8    5  1.003280  11.08089   11.11724   0.07361011
#> 9    6  6.195717  15.41482   95.50588  19.82966088
#> 10   7  1.000115  13.76856   13.77015   0.69467659
#> 11   8  1.000790  10.79428   10.80281   0.00000000
#> 12   9  1.138186  23.72200   27.00004   3.79189947
# }